test(ssh): fold transport-death test into forward_test.go, drop comments
